Spack doesn't run site customization when started + vendor setuptools
fixes #9206
fixes #9034

A possible solution to #9206 is to avoid running `import site` and all
the initialization procedures that this entails. This requires us to use
`python -S` as an interpreter, instead of plain `python`.

Of course things can't be that simple. In Linux (but not in BSD or
MacOS) the shebang pass a single string argument to the interpreter.
More details on the subject are here:

http://sambal.org/2014/02/passing-options-node-shebang-line/
https://github.com/smikes/node/blob/minus-x-switch/doc/Minus-X-Switch-Proposal.md#shebang-interpretation

but the bottom line is that this requires us to have a wrapper bash
script that invokes the interpreter with the correct option.

Finally, as now we are not picking up anything from the 'site', we need
to vendor `setuptools` (which was still missing).